requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
lastpass-python==0.3.2
pycryptodome==3.15.0
requests==2.28.1
urllib3==1.26.12
